Conversion (#7654)
* XCode 15 conversion warning -> error; bypass

* Toss an xcode build in the pipeline also to help

+++ src/surge-xt/gui/widgets/XMLConfiguredMenus.cpp
@@ -322,7 +322,7 @@ void XMLMenuPopulator::populate()
 
     rootTree->buildJuceMenu(menu, this);
 
-    maxIdx = allPresets.size();
+    maxIdx = (int)allPresets.size();
 }
 
 OscillatorMenu::OscillatorMenu() : juce::Component(), WidgetBaseMixin<OscillatorMenu>(this)
